+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/utils/TableReference.java
@@ -26,4 +26,6 @@ public interface TableReference
     String getTable();
 
     String getKeyspace();
+
+    int getGcGraceSeconds();
 }